feat:  시험 문제 답안들 데이터베이스에 저장 구현  - AnswerList.js 파일에 AnswerListSchema, AnswersSchema(서브스키마)를 만듬  - id의 존재여부에 따라 모델을 업데이트  - test/[id] 페이지를 삭제할 때 AnswerList 모델에 id와 [id]가 같으면 제거 기능 구현  - delete 클릭시 confirm 구현  - create 페이지에서 questions type을 number로 변경

+++ src/app/Control.js
@@ -6,21 +6,44 @@ const Control = () => {
     const params = useParams();
     const router = useRouter();
     const id = params.id;
+
+    const handleDelete = async () => {
+        // window.confirm은 확인(true)과 취소(false)가지는 모달창 => 반환 값은 ture or false
+        const confirmDelete = window.confirm('진짜 삭제하시겠습니다?');
+
+        if(!confirmDelete) {
+            return; // 취소 시 아무 작업도 하지 않음
+        }
+
+        const options = {method: 'DELETE'};
+
+        // AnswerList 모델에 삭제할려고하는 id가 존재하면 삭제한다.
+        const answerResponse = await fetch(`${process.env.NEXT_PUBLIC_API_URL}api/answers?id=${id}`, options);
+        const answerDeleteResult = await answerResponse.json();
+        console.log(answerDeleteResult);
+
+        // 현재 [id]를 Topice 모델에서 id값과 일치하는 것을 삭제한다.
+        const topicResponse = await fetch(`${process.env.NEXT_PUBLIC_API_URL}api/topices/${id}`, options);
+        if(topicResponse.ok) {
+            router.push('/') // 홈으로 이동
+            router.refresh(); // 페이지 새로고침
+        } else {
+            console.error('Failed to delete topic');
+        }
+    }
+
     return (
         <ul>
             <li><Link href="/create">Create</Link></li>
             {id ? 
                 <>
-                    <li><input type="button" value='delete' onClick={() => {
-                        const options = {method: 'DELETE'};
-
-                        fetch(`${process.env.NEXT_PUBLIC_API_URL}api/topices/${id}`, options)
-                            .then(response => response.json())
-                            .then(result => {
-                                router.push('/')
-                                router.refresh();
-                            })
-                    }}/></li>
+                    <li>
+                        <input 
+                            type="button"
+                            value='delete'
+                            onClick={handleDelete}
+                        />
+                    </li>
                 </>
             : null}
         </ul>
